Function Calling In Svelte (#258)
For now I am just pushing in the state where I was with my fork of @Zakinator123's fork.  I haven't updated to all the latest comments of that branch, but I wanted to get this available for those that were looking for the Svelte version, if they wanted to clone off this branch to work locally.  I'll upate to include the rest of the changes requested in the other PR here as I can.